src/requirements.txt
azure-functions==1.17.0
azure-functions-durable==1.2.4
azure-servicebus==7.11.4
python-dotenv==1.1.0
